refactor: 도서 대출 로직 Book 도메인이 주도하는 것으로 변경
