+++ src/main/java/com/example/libraryManagement/book/domain/Book.java
@@ -45,6 +45,10 @@ public class Book {
         this.isbn = isbn;
     }
 
+    public boolean isLent() {
+        return this.state.isLent();
+    }
+
     public void lend() {
         this.state = BookState.LENT;
     }
