+++ src/main/java/com/example/libraryManagement/book/domain/BookState.java
@@ -1,5 +1,9 @@
 package com.example.libraryManagement.book.domain;
 
 public enum BookState {
-    LENT, LENDABLE
+    LENT, LENDABLE;
+
+    public boolean isLent() {
+        return this == LENT;
+    }
 }
