+++ src/main/java/com/example/libraryManagement/book/service/BookService.java
@@ -2,11 +2,14 @@ package com.example.libraryManagement.book.service;
 
 import com.example.libraryManagement.book.domain.Book;
 import com.example.libraryManagement.book.domain.BookState;
+import com.example.libraryManagement.book.domain.LentHistory;
 import com.example.libraryManagement.book.dto.request.BookModificationRequest;
 import com.example.libraryManagement.book.dto.request.BookRegisterRequest;
 import com.example.libraryManagement.book.exception.BookException;
 import com.example.libraryManagement.book.service.port.BookRepository;
+import com.example.libraryManagement.book.service.port.LentHistoryRepository;
 import com.example.libraryManagement.book.validator.BookValidator;
+import com.example.libraryManagement.common.response.error.ErrorType;
 import com.example.libraryManagement.user.domain.User;
 import com.example.libraryManagement.user.service.port.UserRepository;
 import lombok.Builder;
@@ -25,6 +28,7 @@ public class BookService {
     private final UserRepository userRepository;
     private final BookRepository bookRepository;
     private final BookValidator bookValidator;
+    private final LentHistoryRepository lentHistoryRepository;
 
     @Transactional
     public Long register(final String registrantEmail, final BookRegisterRequest bookRegisterRequest) {
@@ -60,5 +64,21 @@ public class BookService {
     private Book getBookByBookId(final Long bookId) {
         return bookRepository.findById(bookId)
                 .orElseThrow(() -> new BookException(BOOK_NOT_FOUND));
+    }
+
+    @Transactional
+    public void lend(final String email, final Long bookId) {
+        User user = getUserByEmail(email);
+        Book book = getBookByBookId(bookId);
+        if (book.isLent()) {
+            throw new BookException(ErrorType.BOOK_ALREADY_LENT);
+        }
+        book.lend();
+
+        LentHistory lentHistory = LentHistory.builder()
+                .user(user)
+                .book(book)
+                .build();
+        lentHistoryRepository.save(lentHistory);
     }
 }
