+++ src/main/java/com/example/libraryManagement/book/service/port/LentHistoryRepository.java
@@ -2,6 +2,7 @@ package com.example.libraryManagement.book.service.port;
 
 import com.example.libraryManagement.book.domain.LentHistory;
 
+import java.util.List;
 import java.util.Optional;
 
 public interface LentHistoryRepository {
@@ -9,4 +10,6 @@ public interface LentHistoryRepository {
     LentHistory save(final LentHistory lentHistory);
 
     Optional<LentHistory> findByLentHistoryId(final Long lentHistoryId);
+
+    List<LentHistory> findByBookId(final Long bookId);
 }
