+++ src/test/java/com/example/libraryManagement/book/controller/BookControllerTest.java
@@ -1,6 +1,8 @@
 package com.example.libraryManagement.book.controller;
 
 import com.example.libraryManagement.book.domain.Book;
+import com.example.libraryManagement.book.domain.BookState;
+import com.example.libraryManagement.book.domain.LentHistory;
 import com.example.libraryManagement.book.dto.request.BookModificationRequest;
 import com.example.libraryManagement.book.dto.request.BookRegisterRequest;
 import com.example.libraryManagement.book.exception.BookException;
@@ -10,6 +12,8 @@ import com.example.libraryManagement.user.dto.request.UserRegisterRequest;
 import com.example.libraryManagement.user.exception.UserException;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
+
+import java.util.List;
 
 import static com.example.libraryManagement.common.response.error.ErrorType.*;
 import static org.assertj.core.api.Assertions.*;
@@ -67,5 +71,38 @@ class BookControllerTest {
                 .isbn("10000")
                 .build();
         fakeContainer.bookController.modify(savedBookId, bookModificationRequest, userRegisterRequest.getEmail());
+    }
+
+    @Test
+    @DisplayName("도서 대출 성공")
+    void lend() {
+        FakeContainer fakeContainer = new FakeContainer();
+
+//        유저 생성
+        UserRegisterRequest userRegisterRequest = UserRegisterRequest.builder()
+                .email("a@a.com")
+                .build();
+        fakeContainer.userService.register(userRegisterRequest);
+
+//        도서 생성
+        BookRegisterRequest bookRegisterRequest = BookRegisterRequest.builder()
+                .name("너에게 하고 싶은 말")
+                .isbn("9791191043235")
+                .build();
+        Long savedBookId = fakeContainer.bookService.register(userRegisterRequest.getEmail(), bookRegisterRequest);
+
+//        도서 대출
+        fakeContainer.bookController.lend(savedBookId, userRegisterRequest.getEmail());
+
+//        검증
+        List<LentHistory> lentHistoryList = fakeContainer.lentHistoryRepository.findByBookId(savedBookId);
+        LentHistory lentHistory = lentHistoryList.get(0);
+        User user = fakeContainer.userRepository.findByEmail(userRegisterRequest.getEmail())
+                .orElseThrow(() -> new UserException(USER_NOT_FOUND));
+        Book book = fakeContainer.bookRepository.findById(savedBookId)
+                .orElseThrow(() -> new BookException(BOOK_NOT_FOUND));
+        assertThat(lentHistory.getUser()).isEqualTo(user);
+        assertThat(lentHistory.getBook()).isEqualTo(book);
+        assertThat(book.getState()).isEqualTo(BookState.LENT);
     }
 }
